httpapi/pyproject.toml
[tool.poetry]
name = "bacon-number"
description = "Bacon Number HTTP API"
version = "0.1"
authors = ["Vladimir Ryabtsev"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "=3.6.3"
asynctnt = "=1.2"
asyncpg = "=0.21.0"
networkx = "=2.5"
python-dotenv = "=0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
pytest-asyncio = "^0.14"
